Backend/requirements.txt
# Framework Principal - Compatible con Python 3.11+
fastapi>=0.121.0
uvicorn[standard]>=0.38.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
python-socketio>=5.14.0

# Base de Datos
sqlalchemy>=2.0.44
asyncpg>=0.30.0
alembic>=1.17.0
redis>=7.0.0

# Autenticación y Seguridad
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
slowapi>=0.1.9

# Machine Learning e IA (Opcional - requiere compilador C++ en Windows)
# Para Windows sin Visual Studio, instalar wheels precompilados:
# pip install scikit-learn numpy pandas -f https://download.pytorch.org/whl/torch_stable.html
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
joblib>=1.3.0

# Utilidades
python-dotenv>=1.2.0
email-validator>=2.3.0
httpx>=0.28.0
aiofiles>=25.1.0
aiohttp>=3.13.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.1
locust==2.17.0

# Desarrollo
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoreo
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
